test-project/requirements.txt
requests==2.31.0 # Библиотека для работы с HTTP-запросами
pytest==7.4.0 # Фреймворк для написания и запуска тестов
pytest-html==3.2.0 # Плагин для генерации HTML-отчетов
pytest-xdist==3.3.1 # Плагин для параллельного выполнения тестов
allure-pytest==2.8.18 # Генерация allure отчета